requirements.txt
#REQUIREMENTS FILE

python-dotenv

plotly

gspread

pandas

oauth2client

flask
